third_party/blink/web_tests/platform/win11-arm64/external/wpt/css/css-align/blocks/align-content-block-002-expected.txt

This is a testharness.js-based test.
[FAIL] .test 1: start
  assert_equals: \n<div class="test" style="align-content: start" title="start">\n    <div class="in-flow" data-offset-y="15"></div>\n    <div class="in-flow">\n      <span class="label">START</span>\n      <span class="abspos" data-offset-y="0">ABS</span>\n      <span class="relpos" data-offset-y="0">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 0 but got -1
[FAIL] .test 2: center
  assert_equals: \n<div class="test" style="align-content: center" title="center">\n    <div class="in-flow" data-offset-y="28"></div>\n    <div class="in-flow">\n      <span class="label">CENTER</span>\n      <span class="abspos" data-offset-y="13">ABS</span>\n      <span class="relpos" data-offset-y="13">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 13 but got 12
[FAIL] .test 3: end
  assert_equals: \n<div class="test" style="align-content: end" title="end">\n    <div class="in-flow" data-offset-y="41"></div>\n    <div class="in-flow">\n      <span class="label">END</span>\n      <span class="abspos" data-offset-y="26">ABS</span>\n      <span class="relpos" data-offset-y="26">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 26 but got 25
[FAIL] .test 4: baseline
  assert_equals: \n<div class="test" style="align-content: baseline" title="baseline">\n    <div class="in-flow" data-offset-y="15"></div>\n    <div class="in-flow">\n      <span class="label">BASELINE</span>\n      <span class="abspos" data-offset-y="0">ABS</span>\n      <span class="relpos" data-offset-y="0">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 0 but got -1
[FAIL] .test 5: last baseline
  assert_equals: \n<div class="test" style="align-content: last baseline" title="last baseline">\n    <div class="in-flow" data-offset-y="41"></div>\n    <div class="in-flow">\n      <span class="label">LAST BASELINE</span>\n      <span class="abspos" data-offset-y="26">ABS</span>\n      <span class="relpos" data-offset-y="26">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 41 but got 10
[FAIL] .test 6: flex-start
  assert_equals: \n<div class="test" style="align-content: flex-start" title="flex-start">\n    <div class="in-flow" data-offset-y="15"></div>\n    <div class="in-flow">\n      <span class="label">FLEX-START</span>\n      <span class="abspos" data-offset-y="0">ABS</span>\n      <span class="relpos" data-offset-y="0">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 0 but got -1
[FAIL] .test 7: flex-end
  assert_equals: \n<div class="test" style="align-content: flex-end" title="flex-end">\n    <div class="in-flow" data-offset-y="41"></div>\n    <div class="in-flow">\n      <span class="label">FLEX-END</span>\n      <span class="abspos" data-offset-y="26">ABS</span>\n      <span class="relpos" data-offset-y="26">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 26 but got 25
[FAIL] .test 8: unsafe start
  assert_equals: \n<div class="test" style="align-content: unsafe start" title="unsafe start">\n    <div class="in-flow" data-offset-y="15"></div>\n    <div class="in-flow">\n      <span class="label">UNSAFE START</span>\n      <span class="abspos" data-offset-y="0">ABS</span>\n      <span class="relpos" data-offset-y="0">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 0 but got -1
[FAIL] .test 9: unsafe center
  assert_equals: \n<div class="test" style="align-content: unsafe center" title="unsafe center">\n    <div class="in-flow" data-offset-y="28"></div>\n    <div class="in-flow">\n      <span class="label">UNSAFE CENTER</span>\n      <span class="abspos" data-offset-y="13">ABS</span>\n      <span class="relpos" data-offset-y="13">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 13 but got 12
[FAIL] .test 10: unsafe end
  assert_equals: \n<div class="test" style="align-content: unsafe end" title="unsafe end">\n    <div class="in-flow" data-offset-y="41"></div>\n    <div class="in-flow">\n      <span class="label">UNSAFE END</span>\n      <span class="abspos" data-offset-y="26">ABS</span>\n      <span class="relpos" data-offset-y="26">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 26 but got 25
[FAIL] .test 11: safe start
  assert_equals: \n<div class="test" style="align-content: safe start" title="safe start">\n    <div class="in-flow" data-offset-y="15"></div>\n    <div class="in-flow">\n      <span class="label">SAFE START</span>\n      <span class="abspos" data-offset-y="0">ABS</span>\n      <span class="relpos" data-offset-y="0">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 0 but got -1
[FAIL] .test 12: safe center
  assert_equals: \n<div class="test" style="align-content: safe center" title="safe center">\n    <div class="in-flow" data-offset-y="28"></div>\n    <div class="in-flow">\n      <span class="label">SAFE CENTER</span>\n      <span class="abspos" data-offset-y="13">ABS</span>\n      <span class="relpos" data-offset-y="13">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 13 but got 12
[FAIL] .test 13: safe end
  assert_equals: \n<div class="test" style="align-content: safe end" title="safe end">\n    <div class="in-flow" data-offset-y="41"></div>\n    <div class="in-flow">\n      <span class="label">SAFE END</span>\n      <span class="abspos" data-offset-y="26">ABS</span>\n      <span class="relpos" data-offset-y="26">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 26 but got 25
[FAIL] .test 14: space-evenly
  assert_equals: \n<div class="test" style="align-content: space-evenly" title="space-evenly">\n    <div class="in-flow" data-offset-y="28"></div>\n    <div class="in-flow">\n      <span class="label">SPACE-EVENLY</span>\n      <span class="abspos" data-offset-y="13">ABS</span>\n      <span class="relpos" data-offset-y="13">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 13 but got 12
[FAIL] .test 15: space-between
  assert_equals: \n<div class="test" style="align-content: space-between" title="space-between">\n    <div class="in-flow" data-offset-y="15"></div>\n    <div class="in-flow">\n      <span class="label">SPACE-BETWEEN</span>\n      <span class="abspos" data-offset-y="0">ABS</span>\n      <span class="relpos" data-offset-y="0">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 0 but got -1
[FAIL] .test 16: space-around
  assert_equals: \n<div class="test" style="align-content: space-around" title="space-around">\n    <div class="in-flow" data-offset-y="28"></div>\n    <div class="in-flow">\n      <span class="label">SPACE-AROUND</span>\n      <span class="abspos" data-offset-y="13">ABS</span>\n      <span class="relpos" data-offset-y="13">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected 13 but got 12
[FAIL] .test 17: normal
  assert_equals: \n<div class="test" style="align-content: normal" title="normal">\n    <div class="in-flow" data-offset-y="10"></div>\n    <div class="in-flow">\n      <span class="label">NORMAL</span>\n      <span class="abspos" data-offset-y="-5">ABS</span>\n      <span class="relpos" data-offset-y="-5">REL</span>\n      <div class="overflow">OVERFLOW</div>\n    </div>\n  </div>\noffsetTop expected -5 but got -6
Harness: the test ran to completion.
